+++ types/config.ts
@@ -55,7 +55,7 @@ export interface RootItem {
   completed_notes_count: Record<string, number>
   details: string
   link: string
-  created_at: number
-  updated_at: number
-  days_since_birth: number
+  created_at?: number
+  updated_at?: number
+  days_since_birth?: number
 }